TaxisService and StatisticsService implementation

+++ src/main/java/beans/AverageStatistics.java
@@ -5,33 +5,33 @@ import javax.xml.bind.annotation.XmlRootElement;
 @XmlRootElement
 public class AverageStatistics {
 
-    private int travelledKms;
-    private int batteryLevel;
+    private double travelledKms;
+    private double batteryLevel;
     private double pollutionLevel;
-    private int accomplishedRides;
+    private double accomplishedRides;
 
     public AverageStatistics() {}
 
-    public AverageStatistics(int travelledKms, int batteryLevel, double pollutionLevel, int accomplishedRides) {
+    public AverageStatistics(double travelledKms, double batteryLevel, double pollutionLevel, double accomplishedRides) {
         this.travelledKms = travelledKms;
         this.batteryLevel = batteryLevel;
         this.pollutionLevel = pollutionLevel;
         this.accomplishedRides = accomplishedRides;
     }
 
-    public int getTravelledKms() {
+    public double getTravelledKms() {
         return travelledKms;
     }
 
-    public void setTravelledKms(int travelledKms) {
+    public void setTravelledKms(double travelledKms) {
         this.travelledKms = travelledKms;
     }
 
-    public int getBatteryLevel() {
+    public double getBatteryLevel() {
         return batteryLevel;
     }
 
-    public void setBatteryLevel(int batteryLevel) {
+    public void setBatteryLevel(double batteryLevel) {
         this.batteryLevel = batteryLevel;
     }
 
@@ -43,11 +43,11 @@ public class AverageStatistics {
         this.pollutionLevel = pollutionLevel;
     }
 
-    public int getAccomplishedRides() {
+    public double getAccomplishedRides() {
         return accomplishedRides;
     }
 
-    public void setAccomplishedRides(int accomplishedRides) {
+    public void setAccomplishedRides(double accomplishedRides) {
         this.accomplishedRides = accomplishedRides;
     }
 }
